add flask migrate

+++ ressources/user.py
@@ -1,21 +1,35 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from passlib.hash import pbkdf2_sha256
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import (
+    create_access_token,
+    create_refresh_token,
+    get_jwt_identity,
+    jwt_required,
+    get_jwt,
+)
 
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from models import UserModel
 from schemas import UserSchema
 
+from blocklist import BLOCKLIST
+
 blp = Blueprint("users", __name__, description="Operations on users")
 
 
 @blp.route("/register/")
 class UserRegister(MethodView):
+    @jwt_required(fresh=True)
     @blp.arguments(UserSchema)
+    @blp.doc(security=[{"jwt": []}])
     def post(self, user_data):
-        """register a user"""
+        """register a user if logged in user is admin"""
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         if UserModel.query.filter(UserModel.username == user_data["username"]).first():
             abort(409, message=" A user with that username already exists.")
 
@@ -37,8 +51,14 @@ class User(MethodView):
         user = UserModel.query.get_or_404(user_id)
         return user
 
+    @jwt_required(fresh=True)
+    @blp.doc(security=[{"jwt": []}])
     def delete(self, user_id):
-        """delete user based on user_id"""
+        """delete user based on user_id if logged in user is admin"""
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         user = UserModel.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
@@ -55,6 +75,28 @@ class UserLogin(MethodView):
         ).first()
 
         if user and pbkdf2_sha256.verify(user_data["password"], user.password):
-            access_token = create_access_token(identity=user.id)
-            return {"access_token": access_token}
+            access_token = create_access_token(identity=user.id, fresh=True)
+            refresh_token = create_refresh_token(identity=user.id)
+            return {"access_token": access_token, "refresh_token": refresh_token}
         abort(401, message="Invalid credentials.")
+
+
+@blp.route("/refresh/")
+class TokenRefresh(MethodView):
+    @jwt_required(refresh=True)
+    @blp.doc(security=[{"jwt": []}])
+    def post(self):
+        current_user = get_jwt_identity()
+        new_token = create_access_token(identity=current_user, fresh=False)
+        return {"access_token": new_token}
+
+
+@blp.route("/logout/")
+class UserLogout(MethodView):
+    @jwt_required()
+    @blp.doc(security=[{"jwt": []}])
+    def post(sef):
+        """func that grab the jti and add it to the blocklist"""
+        jti = get_jwt()["jti"]
+        BLOCKLIST.add(jti)
+        return {"message": "Successfully logged out"}
